feat: Added support for runtime aware cache in HybridObject
To make sure jsi objects are deleted and destroyed on the correct runtime in the correct thread, this PR adds a RuntimeAwareCache mechanism that we also use in RNWKC to make sure cached objects are kept alive and only destroyed on the thread that they were created on.

This PR solves this by:
- Adding the RuntimeCache.h file with an adaption of the RuntimeAwareCache system in RNWKC
- Added setup of the cache by passing the main runtime from the FilamentInstaller on Adnroid and in the AppleFilamentProxy.mm on iOS.
- Changed the _functionCache in HybridObject to be a RuntimeAwareCache

NOTE: Check out the pattern of destroying stuff in the dispatcher - it causes crashes especially on reload and might not be needed now.

It should now be possible to reload and run GC without issues.