fix: android release mode assets (#263)
Android release mode bundles `glb` assets into the `raw` folder if they
are bundled with the `require` syntax, instead of linking them with
something like `react-native-assets`. This avoids having to bundle the
assets twice if you are using them via `require`, which works correctly
on iOS.

Otherwise the asset cannot be found by android. And this is a problem
only in release mode Android.